New feature: Allow background color or image slideshows
1. Create a folder "qlock.backgrounds" in userdata/addon_data
2. Add one or more images into this new folder

If you add just one image, this will be used as the background image for
the qlock screensaver. The QLock itself will then appear as a
transparent overlay over the image. This can also be used to set the
background color of QLock: just put a small (1x1 or so) PNG file with
the desired background color into the folder.
If you add multiple images, QLock will play them as a slideshow,
changing once per minute (timeperimage is set to 10000, which would mean
10 seconds - however kodi does read it as every 60 seconds...don't ask
me why)