New feature: background image/slideshow
Use the new "folder" setting of the qlock screensaver add-on if you want
to use it with background images.
If your folder contains just one image, this will be used as the
background image for the qlock screensaver. The QLock itself will then
appear as a transparent overlay over the image. This can also be used to
set the background color of QLock: just put a small (1x1 or so) PNG file
with
the desired background color into the folder.
If you add multiple images, QLock will play them as a slideshow,
changing once per minute.